+++ CHANGELOG.md
@@ -12,6 +12,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 ### Added
 
 - Add flag `--json` to `print-schema` to output JSON instead of GraphQL SDL https://github.com/nuwave/lighthouse/pull/1268
+- Add TTL option for subscriptions storage https://github.com/nuwave/lighthouse/pull/1284
 
 ### Changed